New Feature
Added .env.example and update .gitignore to ignore .env and data.
Also adding the python code for interacting with Ollama model.
There is provision for interacting with Open AI as well. But
this feature is commented. The user will need to uncomment
this, if they wish to use the same.

Worked on the third Issue